chore: separate constants by domain object
